+++ Object.cpp
@@ -4,10 +4,10 @@ Object::Object() :
 	position(Zero),
 	rotation(Zero),
 	velocity(Zero),
-	size(1.0, 1.0, 1.0),
+	size(1.0f, 1.0f, 1.0f),
 	hasCollider(true) {}
 
-void Object::setPosition(GLdouble x, GLdouble y, GLdouble z)
+void Object::setPosition(GLfloat x, GLfloat y, GLfloat z)
 {
 	position = Vec3(x, y, z);
 }
@@ -17,7 +17,7 @@ void Object::setPosition(Vec3 pos)
 	position = pos;
 }
 
-void Object::setRotation(GLdouble x, GLdouble y, GLdouble z)
+void Object::setRotation(GLfloat x, GLfloat y, GLfloat z)
 {
 	rotation = Vec3(x, y, z);
 }
@@ -27,7 +27,7 @@ void Object::setRotation(Vec3 rot)
 	rotation = rot;
 }
 
-void Object::setVelocity(GLdouble x, GLdouble y, GLdouble z)
+void Object::setVelocity(GLfloat x, GLfloat y, GLfloat z)
 {
 	velocity = Vec3(x, y, z);
 }
@@ -37,7 +37,7 @@ void Object::setVelocity(Vec3 v)
 	velocity = v;
 }
 
-void Object::setSize(GLdouble x, GLdouble y, GLdouble z)
+void Object::setSize(GLfloat x, GLfloat y, GLfloat z)
 {
 	size = Vec3(x, y, z);
 }
@@ -47,7 +47,7 @@ void Object::setSize(Vec3 s)
 	size = s;
 }
 
-void Object::applyForce(GLdouble fx, GLdouble fy, GLdouble fz)
+void Object::applyForce(GLfloat fx, GLfloat fy, GLfloat fz)
 {
 	velocity += Vec3(fx, fy, fz);
 }
@@ -57,7 +57,7 @@ void Object::applyForce(Vec3 force)
 	velocity += force;
 }
 
-void Object::updatePosition(GLdouble deltaTime)
+void Object::updatePosition(GLfloat deltaTime)
 {
 	position += velocity * deltaTime;
 }
